Preprocessing the data
Now In the file I have extended the code by preprocessing the data 
In the code I have reorganize the data to include the landmark data and the phrase it represents in each parquet file for efficiency and ease. In this manner, we may avoid alternating between the train.csv and its parquet file.

The updated information will be stored in TFRecord format. A straightforward format for storing a series of binary records is called TFRecord. Using TFRecord to store and load the data is significantly faster and more efficient.

To achieve preprocessing following things have been done 
1. Fetching the pose landmark coordinates related to hand movement.
2.Creating x,y,z label names from coordinates
3. Creating the feature columns from the extracted coordinates.
4. Storing ids of each coordinate labels to lists